+++ src/common/i18n/resources/pt.ts
@@ -361,6 +361,25 @@ const pt = {
       pingFrequencyIsRequired: 'Frequência de verificação é obrigatório.',
     },
   },
+  toasts: {
+    HighResponseTime: {
+      title: ' com ping algo.',
+      message: 'O servidor está com um tempo de resposta muito alto.',
+    },
+    OperatingCorrectly: {
+      title: ' operando normalmente.',
+      message:
+        'O servidor voltou a operar dentro do tempo de resposta correto.',
+    },
+    ServerAvailable: {
+      title: ' disponível novamente.',
+      message: 'O servidor voltou a operar novamente.',
+    },
+    ServerUnavailable: {
+      title: ' indisponível.',
+      message: 'O servidor aparenta estar offline. Verifique.',
+    },
+  },
 };
 
 export default pt;